refactor: 로그인 회원가입 페이지의 타입 types 디렉토리로 옮김
