+++ src/pages/[group-name]/index.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 import { Flex, Heading } from '@chakra-ui/react';
 import GNBLayout from '@/layouts/gnb-layout';
-import MemberListSection from '@/components/pages/member/member-list-section';
-import DetailSection from '@/components/pages/member/detail-section';
+// import MemberListSection from '@/components/pages/member/member-list-section';
+// import DetailSection from '@/components/pages/member/detail-section';
 import { ProfileInfo } from '@/components/common/profile-info/types';
-import CardSendSection from '@/components/pages/member/card-send-section';
+// import CardSendSection from '@/components/pages/member/card-send-section';
 
 export default function Member() {
   const [step, setStep] = useState(0);
@@ -26,13 +26,13 @@ export default function Member() {
         팀원들에게 카드를 보내보세요!
       </Heading>
       <Flex height="full" gap="16px" w="full" overflowY="auto">
-        <MemberListSection
+        {/* <MemberListSection
           isSelected={!!selectedMember}
           onSelect={handleMemberSelect}
-        />
+        /> */}
         {selectedMember && (
           <>
-            {step === 1 && (
+            {/* {step === 1 && (
               <DetailSection
                 selectedMember={selectedMember}
                 onClose={handleClose}
@@ -45,7 +45,7 @@ export default function Member() {
                 onClose={handleClose}
                 onClickCancel={() => setStep(1)}
               />
-            )}
+            )} */}
           </>
         )}
       </Flex>
